requirements.txt
azure-cognitiveservices-speech
python-dotenv
openai
requests
pandas
scipy
streamlit
